requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
pillow==10.3.0
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.2
sqlparse==0.5.0
